存在確認

+++ app/controllers/rooms_controller.rb
@@ -18,8 +18,10 @@ class RoomsController < ApplicationController
 
   def destroy
     room = Room.find(params[:id])
-    room.destroy
-    redirect_to pages_top_path
+    if room.present?
+      room.destroy
+      redirect_to pages_top_path
+    end
   end
 
   private
